Add density curve

Add density curve

Usage

add_density_curve(
  plot,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  color = "#E37D46",
  fill = "#E37D46",
  alpha = 0.3,
  ...
)

Arguments

plot

A tidyplot generated with the function tidyplot().

bw

The smoothing bandwidth to be used. If numeric, the standard deviation of the smoothing kernel. If character, a rule to choose the bandwidth, as listed in stats::bw.nrd(). Note that automatic calculation of the bandwidth does not take weights into account.

adjust

A multiplicate bandwidth adjustment. This makes it possible to adjust the bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half of the default bandwidth.

kernel

Kernel. See list of available kernels in density().

n

number of equally spaced points at which the density is to be estimated, should be a power of two, see density() for details

color

common

fill

common

alpha

common

...

Arguments passed on to the geom function.

Value

A tidyplot object