Continuous color schemes

For more information about the use of color schemes in tidyplots, check out this article: Color schemes

Usage

colors_continuous_viridis

colors_continuous_magma

colors_continuous_inferno

colors_continuous_plasma

colors_continuous_cividis

colors_continuous_rocket

colors_continuous_mako

colors_continuous_turbo

colors_continuous_bluepinkyellow

Format

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 265.

An object of class tidycolor (inherits from character) of length 11.

Details

Color schemes can be conveniently previewed by using the print method of the tidycolor class. This will send a html preview to the RStudio Viewer pane.

colors_continuous_viridis

colors_continuous_viridis
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#440154FF","#460A5DFF","#471264FF","#481B6DFF","#482374FF","#472C7AFF","#46337FFF","#443A83FF","#424186FF","#3F4889FF","#3C508BFF","#39568CFF","#365D8DFF","#33638DFF","#306A8EFF","#2D708EFF","#2B758EFF","#297B8EFF","#26818EFF","#24878EFF","#228D8DFF","#20928CFF","#1F988BFF","#1F9F88FF","#20A486FF","#24AA83FF","#29AF7FFF","#31B57BFF","#3BBB75FF","#45C06FFF","#53C569FF","#5EC962FF","#6ECE58FF","#7BD250FF","#8AD647FF","#9CD93CFF","#AADC32FF","#BDDF26FF","#CCE11EFF","#DEE318FF","#EDE51BFF","#FDE725FF")

colors_continuous_magma

colors_continuous_magma
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#000004FF","#02020DFF","#060519FF","#0C0927FF","#130D34FF","#1C1044FF","#231252FF","#2E1162FF","#38106CFF","#420F75FF","#4E117BFF","#57157EFF","#611880FF","#6A1C81FF","#752181FF","#7D2482FF","#882781FF","#922B81FF","#9B2E7FFF","#A6317DFF","#AF357BFF","#BB3978FF","#C43C75FF","#CF4070FF","#D8456CFF","#DF4B68FF","#E85362FF","#EE5B5EFF","#F4675CFF","#F7715CFF","#FA7E5EFF","#FC8961FF","#FD9668FF","#FEA16EFF","#FEAB75FF","#FEB87EFF","#FEC287FF","#FECF92FF","#FDD99BFF","#FDE6A8FF","#FCF0B2FF","#FCFDBFFF")

colors_continuous_inferno

colors_continuous_inferno
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#000004FF","#02020EFF","#07051AFF","#0D082AFF","#150B37FF","#1F0C48FF","#290B54FF","#350A60FF","#3E0966FF","#490B6AFF","#540F6DFF","#5D126EFF","#68166EFF","#71196EFF","#7C1D6DFF","#85216BFF","#8F2568FF","#992766FF","#A22B62FF","#AD305DFF","#B53458FF","#C03952FF","#C73E4CFF","#D04545FF","#D84C3EFF","#DF5237FF","#E55C30FF","#EB6429FF","#F06F20FF","#F47918FF","#F7840FFF","#F98E09FF","#FB9B06FF","#FCA60CFF","#FCB115FF","#FBBF24FF","#F9C932FF","#F6D746FF","#F3E259FF","#F1EE73FF","#F3F68BFF","#FCFFA4FF")

colors_continuous_plasma

colors_continuous_plasma
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#0D0887FF","#1C068EFF","#290593FF","#360498FF","#3F049CFF","#4B03A1FF","#5502A4FF","#5F01A6FF","#6700A8FF","#7100A8FF","#7B02A8FF","#8405A7FF","#8E0BA5FF","#9511A1FF","#9E199DFF","#A62098FF","#AD2792FF","#B42E8DFF","#BB3488FF","#C23C81FF","#C8437BFF","#CD4A76FF","#D35171FF","#D9586AFF","#DE5F65FF","#E26560FF","#E76E5BFF","#EB7556FF","#EF7E50FF","#F3854BFF","#F68E44FF","#F89540FF","#FB9F3AFF","#FCA835FF","#FDB030FF","#FEBA2CFF","#FDC328FF","#FCCE25FF","#FAD824FF","#F7E325FF","#F4ED27FF","#F0F921FF")

colors_continuous_cividis

colors_continuous_cividis
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#00204DFF","#002557FF","#002961FF","#002D6DFF","#00306FFF","#01366EFF","#173A6DFF","#253E6CFF","#2E436CFF","#36476BFF","#3F4C6BFF","#45506BFF","#4C546CFF","#52596CFF","#595E6DFF","#5E626EFF","#64666FFF","#696B71FF","#6E6F73FF","#747475FF","#797977FF","#7F7D78FF","#848279FF","#8B8779FF","#918C78FF","#979178FF","#9D9677FF","#A39A76FF","#ABA074FF","#B1A573FF","#B7AA71FF","#BEAF6FFF","#C5B56CFF","#CCBB69FF","#D2C066FF","#D9C562FF","#E0CB5EFF","#E8D259FF","#EED753FF","#F7DD4DFF","#FDE346FF","#FFEA46FF")

colors_continuous_rocket

colors_continuous_rocket
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#03051AFF","#0A091FFF","#120D25FF","#1C112BFF","#241432FF","#2E1739FF","#36193FFF","#411B44FF","#491D49FF","#531E4DFF","#5E1F52FF","#671F55FF","#721F57FF","#7B1F59FF","#871E5BFF","#921C5BFF","#9D1B5BFF","#A7195AFF","#B01759FF","#BC1656FF","#C51852FF","#CE1D4EFF","#D62449FF","#DE2E44FF","#E43841FF","#E8413EFF","#ED4F3EFF","#EF5A41FF","#F26747FF","#F3724EFF","#F47E57FF","#F58860FF","#F5946BFF","#F69D75FF","#F6A77FFF","#F6B28CFF","#F6BB97FF","#F7C6A6FF","#F7CEB2FF","#F8D8C1FF","#F9E0CEFF","#FAEBDDFF")

colors_continuous_mako

colors_continuous_mako
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#0B0405FF","#12080DFF","#180D16FF","#1E111FFF","#241628FF","#2A1B33FF","#2F1F3DFF","#342547FF","#372852FF","#3B2D5BFF","#3E3367FF","#403871FF","#413E7EFF","#414387FF","#3F4B90FF","#3D5296FF","#3A5A9AFF","#38619DFF","#37689FFF","#3670A0FF","#3576A2FF","#357DA3FF","#3484A5FF","#348CA7FF","#3492A8FF","#3499AAFF","#35A0ABFF","#37A6ACFF","#3AAEADFF","#3FB5ADFF","#44BDADFF","#4BC2ADFF","#55CAADFF","#61CFACFF","#72D4ADFF","#86D9B1FF","#96DDB5FF","#A9E1BDFF","#B6E6C5FF","#C5EAD0FF","#D0EFDAFF","#DEF5E5FF")

colors_continuous_turbo

colors_continuous_turbo
A tidyplots color scheme with 265 colors, downsampled to 42 colors.

c(
"#30123BFF","#372365FF","#3D3489FF","#4147AEFF","#4456C8FF","#4669E0FF","#4777EFFF","#4688FBFF","#4197FFFF","#38A5FBFF","#2CB7F0FF","#22C4E3FF","#1AD3D1FF","#18DDC2FF","#1DE7B2FF","#29EFA2FF","#3DF58CFF","#53FA79FF","#69FD66FF","#83FF51FF","#98FE43FF","#AAFB39FF","#BAF635FF","#CBED34FF","#D9E436FF","#E4DA38FF","#F0CC3AFF","#F7C13AFF","#FCB136FF","#FEA230FF","#FE8F28FF","#FB7D21FF","#F56918FF","#EF5911FF","#E74A0CFF","#DD3C08FF","#D23105FF","#C32503FF","#B51C01FF","#A31301FF","#910B01FF","#7A0403FF")

colors_continuous_bluepinkyellow

colors_continuous_bluepinkyellow
A tidyplots color scheme with 11 colors.

c(
"#00034D","#000F9F","#001CEF","#241EF5","#5823F6","#A033E0","#E85AB1","#F1907C","#F4AF63","#FCE552","#FFFB6D")