Discrete color schemes

For more information about the use of color schemes in tidyplots, check out this article: Color schemes

Usage

colors_discrete_friendly

colors_discrete_seaside

colors_discrete_apple

colors_discrete_friendly_long

colors_discrete_okabeito

colors_discrete_ibm

colors_discrete_metro

colors_discrete_candy

colors_discrete_alger

colors_discrete_rainbow

Format

An object of class tidycolor (inherits from character) of length 6.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 7.

An object of class tidycolor (inherits from character) of length 7.

An object of class tidycolor (inherits from character) of length 7.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 9.

Details

The signature theme of tidyplots colors_discrete_friendly was adapted from the Okabe & Ito color palette that was designed to work well for people with color vision deficiency.

Color schemes can be conveniently previewed by using the print method of the tidycolor class. This will send a html preview to the RStudio Viewer pane.

colors_discrete_friendly

colors_discrete_friendly
A tidyplots color scheme with 6 colors.

c(
"#0072B2","#56B4E9","#009E73","#F5C710","#E69F00","#D55E00")

colors_discrete_seaside

colors_discrete_seaside
A tidyplots color scheme with 5 colors.

c(
"#8ecae6","#219ebc","#023047","#ffb703","#fb8500")

colors_discrete_apple

colors_discrete_apple
A tidyplots color scheme with 7 colors.

c(
"#ff3b30","#ff9500","#ffcc00","#4cd964","#5ac8fa","#007aff","#5856d6")

colors_discrete_friendly_long

colors_discrete_friendly_long
A tidyplots color scheme with 7 colors.

c(
"#CC79A7","#0072B2","#56B4E9","#009E73","#F5C710","#E69F00","#D55E00")

colors_discrete_okabeito

colors_discrete_okabeito
A tidyplots color scheme with 7 colors.

c(
"#E69F00","#56B4E9","#009E73","#F0E442","#0072B2","#D55E00","#CC79A7")

colors_discrete_ibm

colors_discrete_ibm
A tidyplots color scheme with 5 colors.

c(
"#5B8DFE","#725DEE","#DD227D","#FE5F00","#FFB109")

colors_discrete_metro

colors_discrete_metro
A tidyplots color scheme with 5 colors.

c(
"#4DACD6","#4FAE62","#F6C54D","#E37D46","#C02D45")

colors_discrete_candy

colors_discrete_candy
A tidyplots color scheme with 5 colors.

c(
"#9b5de5","#f15bb5","#fee440","#00bbf9","#00f5d4")

colors_discrete_alger

colors_discrete_alger
A tidyplots color scheme with 5 colors.

c(
"#000000","#1A5B5B","#ACC8BE","#F4AB5C","#D1422F")

colors_discrete_rainbow

colors_discrete_rainbow
A tidyplots color scheme with 9 colors.

c(
"#FF7777","#FFAB74","#FFE577","#DBF47B","#91E480","#7CC9E5","#7DA8E6","#887DE6","#BC7BE4")