Discrete color schemes

Discrete color schemes

Usage

colors_discrete_metro

colors_discrete_circle

colors_discrete_candy

colors_discrete_pastel

colors_discrete_seaside

Format

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 6.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 5.

An object of class tidycolor (inherits from character) of length 5.

Details

Color schemes can be conveniently previewed by using the print method of the tidycolor class. This will send a html preview to the RStudio Viewer pane that looks like this:

colors_discrete_metro
A tidyplots color scheme with 5 colors.

c(
"#4DACD6","#4FAE62","#F6C54D","#E37D46","#C02D45")

colors_discrete_circle
A tidyplots color scheme with 6 colors.

c(
"#939393","#644296","#F08533","#D1352C","#559E3F","#3B78B0")

colors_discrete_candy
A tidyplots color scheme with 5 colors.

c(
"#9b5de5","#f15bb5","#fee440","#00bbf9","#00f5d4")

colors_discrete_pastel
A tidyplots color scheme with 5 colors.

c(
"#cdb4db","#ffc8dd","#ffafcc","#bde0fe","#a2d2ff")

colors_discrete_seaside
A tidyplots color scheme with 5 colors.

c(
"#8ecae6","#219ebc","#023047","#ffb703","#fb8500")