Diverging color schemes

For more information about the use of color schemes in tidyplots, check out this article: Color schemes

Usage

colors_diverging_blue2brown

colors_diverging_blue2red

colors_diverging_BuRd

colors_diverging_BuYlRd

Format

An object of class tidycolor (inherits from character) of length 17.

An object of class tidycolor (inherits from character) of length 9.

An object of class tidycolor (inherits from character) of length 9.

An object of class tidycolor (inherits from character) of length 9.

Details

Color schemes can be conveniently previewed by using the print method of the tidycolor class. This will send a html preview to the RStudio Viewer pane that looks like this:

colors_diverging_blue2brown
A tidyplots color scheme with 17 colors.

c(
"#1961A5","#2671B5","#2D80BF","#268CC9","#119DD8","#00B2EB","#66C5EF","#C4E5F8","#FEFCF6","#FDEEB8","#FCDD67","#F6C445","#E78B43","#DD5642","#DB3E34","#CA3632","#B3322E")

colors_diverging_blue2red
A tidyplots color scheme with 9 colors.

c(
"#0000FF","#3F3FFF","#7F7FFF","#BFBFFF","#FFFFFF","#FFBFBF","#FF7F7F","#FF3F3F","#FF0000")

colors_diverging_BuRd
A tidyplots color scheme with 9 colors.

c(
"#2166AC","#4393C3","#92C5DE","#D1E5F0","#F7F7F7","#FDDBC7","#F4A582","#D6604D","#B2182B")

colors_diverging_BuYlRd
A tidyplots color scheme with 9 colors.

c(
"#4575B4","#74ADD1","#ABD9E9","#E0F3F8","#FFFFBF","#FEE090","#FDAE61","#F46D43","#D73027")