Diverging color schemes

For more information about the use of color schemes in tidyplots, check out this article: Color schemes

Usage

colors_diverging_blue2red

colors_diverging_blue2brown

colors_diverging_BuRd

colors_diverging_BuYlRd

colors_diverging_spectral

colors_diverging_icefire

Format

An object of class tidycolor (inherits from character) of length 17.

An object of class tidycolor (inherits from character) of length 17.

An object of class tidycolor (inherits from character) of length 11.

An object of class tidycolor (inherits from character) of length 11.

An object of class tidycolor (inherits from character) of length 96.

An object of class tidycolor (inherits from character) of length 96.

Details

Color schemes can be conveniently previewed by using the print method of the tidycolor class. This will send a html preview to the RStudio Viewer pane.

colors_diverging_blue2red

colors_diverging_blue2red
A tidyplots color scheme with 17 colors.

c(
"#0000FF","#1F1FFF","#3F3FFF","#5F5FFF","#7F7FFF","#9F9FFF","#BFBFFF","#DFDFFF","#FFFFFF","#FFDFDF","#FFBFBF","#FF9F9F","#FF7F7F","#FF5F5F","#FF3F3F","#FF1F1F","#FF0000")

colors_diverging_blue2brown

colors_diverging_blue2brown
A tidyplots color scheme with 17 colors.

c(
"#1961A5","#2671B5","#2D80BF","#268CC9","#119DD8","#00B2EB","#66C5EF","#C4E5F8","#FEFCF6","#FDEEB8","#FCDD67","#F6C445","#E78B43","#DD5642","#DB3E34","#CA3632","#B3322E")

colors_diverging_BuRd

colors_diverging_BuRd
A tidyplots color scheme with 11 colors.

c(
"#053061","#2166AC","#4393C3","#92C5DE","#D1E5F0","#F7F7F7","#FDDBC7","#F4A582","#D6604D","#B2182B","#67001F")

colors_diverging_BuYlRd

colors_diverging_BuYlRd
A tidyplots color scheme with 11 colors.

c(
"#313695","#4575B4","#74ADD1","#ABD9E9","#E0F3F8","#FFFFBF","#FEE090","#FDAE61","#F46D43","#D73027","#A50026")

colors_diverging_spectral

colors_diverging_spectral
A tidyplots color scheme with 96 colors, downsampled to 42 colors.

c(
"#5b53a4","#525fa9","#486cb0","#3f77b5","#3389bd","#3d95b8","#47a0b3","#58b2ac","#64c0a6","#71c6a5","#86cfa5","#94d4a4","#a2d9a4","#b5e1a2","#c3e79f","#cfec9d","#e1f399","#e9f69d","#eef8a4","#f6fbb0","#fcfeba","#fffdbc","#fff7b2","#feeda1","#fee797","#fee08b","#fed27f","#fdc776","#fdbd6d","#fdad60","#fba05b","#f99153","#f67f4b","#f47044","#ef6645","#e55749","#df4e4b","#d9444d","#cb334d","#be254a","#b41947","#a90d45")

colors_diverging_icefire

colors_diverging_icefire
A tidyplots color scheme with 96 colors, downsampled to 42 colors.

c(
"#b7e3d9","#a9d9d6","#98cdd2","#8ac4d0","#72b6ce","#63adcd","#55a3cd","#4394ce","#3987cf","#377cd0","#3f69c9","#465ebe","#4954b0","#474792","#42407b","#3c3a69","#323050","#2c2b42","#272636","#212028","#1f1e21","#201e1e","#261e1f","#332023","#3d2228","#4a252e","#5c2935","#6d2b3b","#7b2d40","#932e44","#a22f44","#b33341","#c53c3c","#d24737","#da5334","#e66734","#eb753a","#ef8445","#f39a5f","#f7ab75","#fab887","#fcc69a")