Format numbers or p values

Format numbers or p values

Usage

format_number(x, accuracy = 0.1, big.mark = ",", scale_cut = NULL, ...)

format_p_value(x, accuracy = 1e-04)

Arguments

x

A number to format.

accuracy

A number to round to. Use (e.g.) 0.01 to show 2 decimal places of precision. If NULL, the default, uses a heuristic that should ensure breaks have the minimum number of digits needed to show the difference between adjacent values.

Applied to rescaled data.

big.mark

Character used between every 3 digits to separate thousands.

scale_cut

Named numeric vector that allows you to rescale large (or small) numbers and add a prefix. Built-in helpers include:

cut_short_scale(): [10^3, 10^6) = K, [10^6, 10^9) = M, [10^9, 10^12) = B, [10^12, Inf) = T.

cut_long_scale(): [10^3, 10^6) = K, [10^6, 10^12) = M, [10^12, 10^18) = B, [10^18, Inf) = T.

cut_si(unit): uses standard SI units.

If you supply a vector c(a = 100, b = 1000), absolute values in the range [0, 100) will not be rescaled, absolute values in the range [100, 1000) will be divided by 100 and given the suffix "a", and absolute values in the range [1000, Inf) will be divided by 1000 and given the suffix "b". If the division creates an irrational value (or one with many digits), the cut value below will be tried to see if it improves the look of the final label.

...

Arguments passed on to scales::number

scale

A scaling factor: x will be multiplied by scale before formatting. This is useful if the underlying data is very small or very large.

style_positive

A string that determines the style of positive numbers:

"none" (the default): no change, e.g. 1.

"plus": preceded by +, e.g. +1.

"space": preceded by a Unicode "figure space", i.e., a space equally as wide as a number or +. Compared to "none", adding a figure space can ensure numbers remain properly aligned when they are left- or right-justified.

style_negative

A string that determines the style of negative numbers:

"hyphen" (the default): preceded by a standard hypen -, e.g. -1.

"minus", uses a proper Unicode minus symbol. This is a typographical nicety that ensures - aligns with the horizontal bar of the the horizontal bar of +.

"parens", wrapped in parentheses, e.g. (1).

Value

Formatted number as character string.

Examples

format_number(232342.3443)
#> [1] "232,342.3"

format_number(232342.3443, accuracy = 0.01)
#> [1] "232,342.34"

format_number(232342.3443, accuracy = 1, big.mark = "")
#> [1] "232342"

format_p_value(0.03445553)
#> [1] "0.0345"

format_p_value(0.0003445553)
#> [1] "0.0003"

format_p_value(0.00003445553)
#> [1] "< 0.0001"